Split optional class from webview.h
